+++ contracts/interfaces/IUniswapV2Router.sol
@@ -1,10 +1,12 @@
 pragma solidity >=0.6.0 <0.9.0;
 
 interface IUniswapV2Router {
-    function swapExactETHForTokens(uint amountOutMin, address[] calldata path, address to, uint deadline)
-          external
-          payable
-          returns (uint[] memory amounts);
+    function swapExactETHForTokens(
+        uint256 amountOutMin,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    ) external payable returns (uint256[] memory amounts);
 
-    function getAmountsOut(uint amountIn, address[] calldata path) external view returns (uint[] memory amounts);
+    function getAmountsOut(uint256 amountIn, address[] calldata path) external view returns (uint256[] memory amounts);
 }
